chore: remove legacy job/op code, add README and documentation
