+++ docs/README.md
@@ -1,0 +1,6 @@
+# Documentation
+
+## Asset DAG screenshot
+
+After running `dagster dev`, capture a screenshot of the asset DAG from the
+Dagster UI and save it here as `asset_dag.png`.
