+++ tests/test_processing.py
@@ -1,25 +1,18 @@
-"""Smoke tests for helix_dagster.processing."""
+"""Smoke tests for core modules."""
 
-import os
+import importlib
 
 
-def test_import_processing_with_missing_yaml(monkeypatch):
-    """Verify processing.py can be imported when COORD_TRANSFORMS_YAML points to
-    a nonexistent file — the CoordinateTransformer.from_yaml call may raise, but
-    import-time errors should be catchable, not crash the interpreter."""
-    # Point to a nonexistent YAML so the module-level init hits a missing file
+def test_import_coordinates_with_missing_yaml(monkeypatch):
+    """Verify coordinates.py can be imported when COORD_TRANSFORMS_YAML points to
+    a nonexistent file — the module-level init should handle it gracefully."""
     monkeypatch.setenv("COORD_TRANSFORMS_YAML", "/tmp/nonexistent_transforms.yaml")
 
-    # Reload the module to trigger the module-level CoordinateTransformer init
-    import importlib
-
     try:
-        import helix_dagster.processing as proc_mod
+        import helix_dagster.coordinates as coord_mod
 
-        importlib.reload(proc_mod)
+        importlib.reload(coord_mod)
     except Exception as exc:
-        # It's acceptable for the import to raise (FileNotFoundError, etc.)
-        # as long as it doesn't segfault or produce an unrecoverable error.
         assert isinstance(exc, Exception), f"Unexpected error type: {type(exc)}"
 
 
@@ -28,32 +21,15 @@ def test_constants_importable():
     from helix_dagster.constants import COLUMN_MAP, IGSN_PATTERN
 
     assert "Sample_ID" in COLUMN_MAP
-    # Verify the case fix: should map to "Sample_IGSN" (uppercase S)
     assert COLUMN_MAP["Sample_ID"] == "Sample_IGSN"
     assert IGSN_PATTERN.search("ABCDEF12345") is not None
 
 
-def test_find_pdv_matches():
-    """Test the pure find_pdv_matches function."""
-    from helix_dagster.processing import find_pdv_matches
-
-    pdv_items = [
-        {"name": "shot001_ch1.tdms"},
-        {"name": "shot001_ch2.tdms"},
-        {"name": "shot002_ch1.tdms"},
-    ]
-    assert len(find_pdv_matches(pdv_items, "shot001")) == 2
-    assert len(find_pdv_matches(pdv_items, "shot002")) == 1
-    assert len(find_pdv_matches(pdv_items, "shot003")) == 0
-
-
 def test_nan_to_none():
-    """Test the _nan_to_none helper."""
-    import math
+    """Test the nan_to_none helper."""
+    from helix_dagster.girder_io import nan_to_none
 
-    from helix_dagster.processing import _nan_to_none
-
-    assert _nan_to_none(float("nan")) is None
-    assert _nan_to_none(42) == 42
-    assert _nan_to_none("hello") == "hello"
-    assert _nan_to_none(None) is None
+    assert nan_to_none(float("nan")) is None
+    assert nan_to_none(42) == 42
+    assert nan_to_none("hello") == "hello"
+    assert nan_to_none(None) is None
